chore: simplify import structure
