+++ src/common/core/logger.ts
@@ -1,5 +1,5 @@
-import { ConsoleStream, Level, Logger } from 'optic';
-import { TokenReplacer } from 'optic/formatters';
+import { ConsoleStream, Level, Logger } from '@optic';
+import { TokenReplacer } from '@optic/formatters';
 import { LogtailStream } from '../logger/logtail.ts';
 import { env } from './env.ts';
 import { MinLogLevel } from '../logger/types.ts';
