+++ src/common/core/setup.ts
@@ -1,46 +1,9 @@
 import { State } from '../types/state.ts';
 import { env } from './env.ts';
-import { GrowthBook } from 'growthbook';
+import { GrowthBook } from '@growthbook';
 import { logger } from './logger.ts';
-import { between } from 'optic';
+import { between } from '@optic';
 import _localSourceFactory from '../mongo/factory.ts';
-
-logger.mark('setup-start');
-
-const applicationState: State = {
-  credential: {
-    supabase: {
-      id: env<string>('SUPABASE_ID'),
-      key: env<string>('SUPABASE_API_KEY'),
-    },
-  },
-  features: new GrowthBook({
-    apiHost: env<string>('GROWTH'),
-    clientKey: env<string>('GROWTH_KEY'),
-    enableDevMode: env<boolean>('GROWTH_DEV_MODE'),
-    log: (msg, ctx) => {
-      logger.info(msg, ctx);
-    },
-    trackingCallback: (experiment, result) => {
-      // substitute with segment or something else for exp tracking
-      logger.debug('Experiemnt tracked', {
-        experimentId: experiment.key,
-        variationId: result.key,
-      });
-    },
-    onFeatureUsage: (featureKey, result) => {
-      logger.debug('Feature used', { key: featureKey, value: result.value });
-    },
-  }),
-  contextHeader: {
-    agent: '',
-    accepts: [],
-    authorization: null,
-    contentType: null,
-    acceptEncoding: '',
-  },
-  local: await _localSourceFactory.connect(),
-};
 
 const onDispose = (token: number) => {
   setTimeout(() => {
@@ -66,7 +29,44 @@ const onTerminationRequest = (): void => {
 Deno.addSignalListener('SIGINT', onTerminationRequest);
 Deno.addSignalListener('SIGTERM', onTerminationRequest);
 
-logger.mark('setup-end');
-logger.measure(between('setup-start', 'setup-end'));
+export const createState = async (): Promise<State> => {
+  logger.mark('setup-start');
+  const state = {
+    credential: {
+      supabase: {
+        id: env<string>('SUPABASE_ID'),
+        key: env<string>('SUPABASE_API_KEY'),
+      },
+    },
+    features: new GrowthBook({
+      apiHost: env<string>('GROWTH'),
+      clientKey: env<string>('GROWTH_KEY'),
+      enableDevMode: env<boolean>('GROWTH_DEV_MODE'),
+      log: (msg, ctx) => {
+        logger.info(msg, ctx);
+      },
+      trackingCallback: (experiment, result) => {
+        // substitute with segment or something else for exp tracking
+        logger.debug('Experiemnt tracked', {
+          experimentId: experiment.key,
+          variationId: result.key,
+        });
+      },
+      onFeatureUsage: (featureKey, result) => {
+        logger.debug('Feature used', { key: featureKey, value: result.value });
+      },
+    }),
+    contextHeader: {
+      agent: '',
+      accepts: [],
+      authorization: null,
+      contentType: null,
+      acceptEncoding: '',
+    },
+    local: await _localSourceFactory.connect(),
+  };
 
-export default applicationState;
+  logger.mark('setup-end');
+  logger.measure(between('setup-start', 'setup-end'));
+  return state;
+};
