+++ src/common/middleware/growth.ts
@@ -1,4 +1,4 @@
-import { between } from 'optic';
+import { between } from '@optic';
 import { logger } from '../core/logger.ts';
 import type { AppContext } from '../types/core.ts';
 import { env } from '../core/env.ts';
