+++ src/common/types/core.ts
@@ -1,7 +1,7 @@
-import { Context } from 'oak';
+import { Context } from '@oak';
 import { State } from './state.ts';
-import { GrowthBook } from 'growthbook';
-import { Db } from 'npm/mongodb';
+import { GrowthBook } from '@growthbook';
+import { Db } from '@mongodb';
 import { AppFeatures } from '../experiment/types.ts';
 
 export type RCF822Date = string;
