+++ src/config/index.ts
@@ -2,7 +2,7 @@ import { AppContext, ErrorResponse } from '../common/types/core.ts';
 import { ConfigRepository } from './repository/index.ts';
 import { LocalSource } from './local/index.ts';
 import { collection } from '../common/mongo/index.ts';
-import { Status } from 'oak';
+import { Status } from '@oak';
 
 export const config = async ({ state, response }: AppContext) => {
   const localSource = new LocalSource(collection('config', state.local));
