+++ src/import_map.json
@@ -1,18 +1,18 @@
 {
   "imports": {
-    "std/http": "jsr:@std/http@1.0.10",
-    "std/dotenv": "jsr:@std/dotenv@0.225.2",
-    "std/collections": "jsr:@std/collections@1.0.9",
-    "std/testing": "jsr:@std/testing@1.0.9",
-    "std/testing/bdd": "jsr:@std/testing@1.0.9/bdd",
-    "std/assert": "jsr:@std/assert@1.0.8",
-    "optic": "jsr:@onjara/optic@2.0.3",
-    "optic/formatters": "jsr:@onjara/optic@2.0.3/formatters",
-    "oak": "jsr:@oak/oak@17.1.3",
-    "deepmerge": "jsr:@rebeccastevens/deepmerge@7.1.3",
-    "growthbook": "jsr:@growthbook/growthbook@1.2.2",
-    "xml": "jsr:@libs/xml@6.0.1",
-    "npm/logtail": "npm:@logtail/node@0.5.2",
-    "npm/mongodb": "npm:mongodb@6.13.0"
+    "@std/http": "jsr:@std/http@1.0.10",
+    "@std/dotenv": "jsr:@std/dotenv@0.225.2",
+    "@std/collections": "jsr:@std/collections@1.0.9",
+    "@std/testing": "jsr:@std/testing@1.0.9",
+    "@std/testing/bdd": "jsr:@std/testing@1.0.9/bdd",
+    "@std/assert": "jsr:@std/assert@1.0.8",
+    "@optic": "jsr:@onjara/optic@2.0.3",
+    "@optic/formatters": "jsr:@onjara/optic@2.0.3/formatters",
+    "@oak": "jsr:@oak/oak@17.1.3",
+    "@deepmerge": "jsr:@rebeccastevens/deepmerge@7.1.3",
+    "@growthbook": "jsr:@growthbook/growthbook@1.2.2",
+    "@xml": "jsr:@libs/xml@6.0.1",
+    "@logtail": "npm:@logtail/node@0.5.2",
+    "@mongodb": "npm:mongodb@6.13.0"
   }
 }
