+++ src/news/local/news.local.source.ts
@@ -1,11 +1,11 @@
-import { Collection, Filter, FindOptions, ObjectId, WithId } from 'npm/mongodb';
+import { Collection, Filter, FindOptions, ObjectId, WithId } from '@mongodb';
 import { logger } from '../../common/core/logger.ts';
 import { IPaging } from '../../common/types/paging.ts';
 import { IResponse } from '../../common/types/response.ts';
 import { toDocument, toEntity } from '../mapper/index.ts';
 import { News } from '../types.ts';
 import { NewsDocument, NewsId } from './types.ts';
-import { between } from 'optic';
+import { between } from '@optic';
 import { projectionOf, sortOf } from '../../common/mongo/index.ts';
 
 export default class LocalSource {
