+++ src/series/local/types.ts
@@ -1,4 +1,4 @@
-import { Document } from 'npm/mongodb';
+import { Document } from '@mongodb';
 import { MediaWithSeason } from '../types.ts';
 
 export interface MediaDocument extends Document, MediaWithSeason {
