+++ src/server.ts
@@ -2,6 +2,8 @@ import factory from './common/core/factory.ts';
 import { port } from './common/core/utils.ts';
 import router from './routes.ts';
 
-await factory({
+const instance = await factory({
   router: router,
-}).listen({ port });
+});
+
+instance.listen({ port });
